+++ src/OpenTelemetry.Instrumentation.Process/CHANGELOG.md
@@ -11,6 +11,9 @@
   metric as an alternative.
   ([#4088](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4088))
 
+* Updated OpenTelemetry core component version(s) to `1.16.0`.
+  ([#4487](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4487))
+
 ## 1.15.1-beta.1
 
 Released 2026-Apr-21